# Speed up base64 encoding 2x by using mutating iterators and better preallocation