--- a/benches/base64.rs
+++ b/benches/base64.rs
@@ -27,3 +27,22 @@ fn bench_from_base64(b: &mut Bencher) {
     b.bytes = sb.len() as u64;
 }
 
+
+#[bench]
+fn bench_to_base64_large(b: &mut Bencher) {
+    let s: Vec<_> = (0..10000).map(|i| ((i as u32 * 12345) % 256) as u8).collect();
+    b.iter(|| {
+        s.to_base64(STANDARD);
+    });
+    b.bytes = s.len() as u64;
+}
+
+#[bench]
+fn bench_from_base64_large(b: &mut Bencher) {
+    let s: Vec<_> = (0..10000).map(|i| ((i as u32 * 12345) % 256) as u8).collect();
+    let sb = s.to_base64(STANDARD);
+    b.iter(|| {
+        sb.from_base64().unwrap();
+    });
+    b.bytes = sb.len() as u64;
+}