Iterators are really fast. So is using mutation instead of pushing.

Before:

    test bench_to_base64         ... bench:       356 ns/iter (+/- 10) = 424 MB/s
    test bench_to_base64_large   ... bench:     21378 ns/iter (+/- 1030) = 467 MB/s

After:

    test bench_to_base64         ... bench:       207 ns/iter (+/- 10) = 729 MB/s
    test bench_to_base64_large   ... bench:     10900 ns/iter (+/- 709) = 917 MB/s

[Diff best seen ignoring whitespace.](https://github.com/rust-lang/rustc-serialize/pull/119/files?w=1)